+++ src/backend/metal/src/window.rs
@@ -244,12 +244,6 @@ impl w::PresentationSurface<Backend> for Surface {
         if !image::Usage::COLOR_ATTACHMENT.contains(config.image_usage) {
             warn!("Swapchain usage {:?} is not expected", config.image_usage);
         }
-        #[cfg(target_os = "macos")]
-        {
-            if self.view.is_some() && self.main_thread_id != thread::current().id() {
-                return Err(w::SwapchainError::WrongThread);
-            }
-        }
         self.swapchain_format = self.configure(&device.shared, &config);
         Ok(())
     }